+++ WebApp1.0/Models/Model_mcircleResult.cs
@@ -9,10 +9,10 @@ namespace WebApp1._0.Models
     public class Model_mcircleResult
     {
         [Key]
-        public int circleid { get; set; }
-        public int circlecode { get; set; }
+        public byte circleid { get; set; }
+        public byte circlecode { get; set; }
         public string circlename { get; set; }
-        public int ref_zoneid { get; set; }
+        public byte ref_zoneid { get; set; }
         public string zonename { get; set; }
        
     }
